requirements.txt
streamlit==1.23.1
matplotlib==3.5.2
numpy==1.23.5
opencv-python-headless==4.7.0.72
pandas==1.4.4
seaborn==0.11.2
squarify==0.4.3
pillow==9.2.0
wordcloud==1.8.2.2
plotly==5.9.0
airtable-python-wrapper==0.15.3
cssselect==1.1.0
matplotlib==3.5.2
numpy==1.23.5
opt-einsum==3.3.0
outcome==1.2.0
pyarrow==12.0.0
scipy==1.9.1
selenium==3.141.0
streamlit-folium==0.12.0
webdriver-manager==3.8.6
webencodings==0.5.1
requests==2.28.1
beautifulsoup4==4.11.1
